app/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
easyocr==1.4.2
Flask==2.1.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
loguru==0.6.0
lxml==4.8.0
MarkupSafe==2.1.1
networkx==2.7.1
PyMuPDF==1.19.6
python-docx==0.8.11
requests==2.27.1
retrying==1.3.3
urllib3==1.26.9
Werkzeug==2.1.1
zipp==3.8.0
